+++ src/index.ts
@@ -13,5 +13,5 @@ export type { CertificateIssuanceOptions } from './lib/pki/CertificateIssuanceOp
 export { generateTxtRdata } from './lib/dns/rdataSerialisation.js';
 export { retrieveDnssecChain } from './lib/dns/dnssecChainRetrieval.js';
 export { serialiseMemberIdBundle } from './lib/memberIdBundle/serialisation.js';
-export { sign } from './lib/signature/generation.js';
+export { sign } from './lib/signature.js';
 export type { IDatePeriod } from './lib/dates.js';
